backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
pydantic-settings==2.10.1
redis==6.4.0
langchain==0.3.27
langgraph==0.6.6
pgvector==0.4.1
python-multipart==0.0.20
httpx==0.28.1
pypdf==5.9.0
python-docx==1.2.0
